🧹 refactor: consolidate AI toggle logic into AIToggleManager and integrate with AICommandsManager for shared detection cache
