+++ src/keymaps/index.ts
@@ -3,4 +3,3 @@ export * from './ai/utils';
 export * from './terminal/left-right';
 export * from './terminal/bottom';
 export * from './terminal/shared';
-export * from './ai/toggle-handler';
